Change calculations from f64 to f32
If precision is found to be necessary somewhere it can be added at that point.

+++ src/action_state.rs
@@ -4,15 +4,15 @@ use std::default::{Default};
 
 #[derive(Clone)]
 pub struct ActionState {
-	pub movement_direction: Vector2<f64>,
-	pub view_direction:     Vector2<f64>,
+	pub movement_direction: Vector2<f32>,
+	pub view_direction:     Vector2<f32>,
 }
 
 impl Default for ActionState {
 	fn default() -> ActionState {
 		ActionState {
-			movement_direction: Vector2{ x: 0f64, y: 0f64 },
-			view_direction:     Vector2{ x: 0f64, y: 0f64 },
+			movement_direction: Vector2{ x: 0f32, y: 0f32 },
+			view_direction:     Vector2{ x: 0f32, y: 0f32 },
 		}
 	}
 }
